nvidia_agno_agent/requirements.txt
# Agno Framework and AgentOS
agno>=2.0.0
agent-os>=1.0.0

# Core AI and LLM
openai>=1.0.0
anthropic>=0.20.0

# Vector Database and Embeddings
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Web Search and Data Processing
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
aiohttp>=3.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
toml>=0.10.0

# Database
sqlalchemy>=2.0.0
sqlite3

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Monitoring and Logging
structlog>=23.0.0
rich>=13.0.0
